KIM-209 ♻️ Member와 Profile Service를 분리하다
