+++ src/main/java/com/devcourse/be04daangnmarket/member/application/MemberService.java
@@ -1,13 +1,9 @@
 package com.devcourse.be04daangnmarket.member.application;
 
 import com.devcourse.be04daangnmarket.member.domain.Member;
-import com.devcourse.be04daangnmarket.member.domain.Profile;
 import com.devcourse.be04daangnmarket.member.dto.MemberDto;
-import com.devcourse.be04daangnmarket.member.dto.ProfileDto;
 import com.devcourse.be04daangnmarket.member.repository.MemberRepository;
-import com.devcourse.be04daangnmarket.member.repository.ProfileRepository;
 import com.devcourse.be04daangnmarket.member.util.MemberConverter;
-import com.devcourse.be04daangnmarket.member.util.ProfileConverter;
 import jakarta.transaction.Transactional;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -15,24 +11,19 @@ import org.springframework.stereotype.Service;
 
 import java.util.NoSuchElementException;
 
-import static com.devcourse.be04daangnmarket.member.exception.ErrorMessage.DUPLICATED_USERNAME;
 import static com.devcourse.be04daangnmarket.member.exception.ErrorMessage.FAIL_LOGIN;
 import static com.devcourse.be04daangnmarket.member.exception.ErrorMessage.ILLEGAL_USER_ACCESS;
-import static com.devcourse.be04daangnmarket.member.exception.ErrorMessage.NOT_FOUND_PROFILE;
 import static com.devcourse.be04daangnmarket.member.exception.ErrorMessage.NOT_FOUND_USER;
 
 @Service
 @Transactional
 public class MemberService {
     private final MemberRepository memberRepository;
-    private final ProfileRepository profileRepository;
     private final PasswordEncoder passwordEncoder;
 
     public MemberService(MemberRepository memberRepository,
-                         ProfileRepository profileRepository,
                          PasswordEncoder passwordEncoder) {
         this.memberRepository = memberRepository;
-        this.profileRepository = profileRepository;
         this.passwordEncoder = passwordEncoder;
     }
 
@@ -70,29 +61,6 @@ public class MemberService {
         return MemberConverter.toResponse(member);
     }
 
-    public ProfileDto.Response updateProfile(Long id, String username) {
-        Profile profile = getProfile(id);
-
-        if (isAvailableUsername(username)) {
-            profile.updateProfile(username);
-
-            return ProfileConverter.toResponse(profile);
-        }
-
-        throw new IllegalArgumentException(DUPLICATED_USERNAME.getMessage());
-    }
-
-    public Profile getProfile(Long memberId) {
-        return profileRepository.findByMemberId(memberId)
-                .orElseThrow(() -> new NoSuchElementException(NOT_FOUND_PROFILE.getMessage()));
-    }
-
-    public ProfileDto.Response toProfile(Long id) {
-        Profile profile = getProfile(id);
-
-        return ProfileConverter.toResponse(profile);
-    }
-
     public void validateById(Long pathId, Long authUserId) {
         if (pathId != authUserId) {
             throw new IllegalArgumentException(ILLEGAL_USER_ACCESS.getMessage());
@@ -106,9 +74,5 @@ public class MemberService {
 
     public boolean isExistMember(String email) {
         return memberRepository.findByEmail(email).isPresent();
-    }
-
-    private boolean isAvailableUsername(String username) {
-        return memberRepository.findByUsername(username).isEmpty();
     }
 }
