+++ src/main/java/com/devcourse/be04daangnmarket/member/repository/MemberRepository.java
@@ -9,6 +9,4 @@ import java.util.Optional;
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
     Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByUsername(String username);
 }
