+++ src/main/java/com/devcourse/be04daangnmarket/post/application/PostService.java
@@ -7,6 +7,7 @@ import java.util.List;
 import java.util.NoSuchElementException;
 
 import com.devcourse.be04daangnmarket.image.dto.ImageDto;
+import com.devcourse.be04daangnmarket.member.application.ProfileService;
 import com.devcourse.be04daangnmarket.post.domain.constant.PostStatus;
 import com.devcourse.be04daangnmarket.post.util.PostConverter;
 
@@ -35,12 +36,12 @@ import jakarta.servlet.http.HttpServletResponse;
 public class PostService {
     private final PostRepository postRepository;
     private final ImageService imageService;
-    private final MemberService memberService;
+    private final ProfileService profileService;
 
-    public PostService(PostRepository postRepository, ImageService imageService, MemberService memberService) {
+    public PostService(PostRepository postRepository, ImageService imageService, ProfileService profileService) {
         this.postRepository = postRepository;
         this.imageService = imageService;
-        this.memberService = memberService;
+        this.profileService = profileService;
     }
 
     @Transactional
@@ -194,7 +195,7 @@ public class PostService {
     }
 
     private String getUsername(Long memberId) {
-        return memberService.toProfile(memberId).username();
+        return profileService.toProfile(memberId).username();
     }
 
     private boolean isViewed(Long id, HttpServletRequest req, HttpServletResponse res) {
